tests: move ab() to conftest
