+++ tests/functional/test_buildah.py
@@ -11,29 +11,10 @@ import pytest
 
 from ansible_bender.builders.buildah_builder import buildah, inspect_resource, \
     podman_run_cmd
-from ..spellbook import basic_playbook_path, base_image, project_dir, \
-    bad_playbook_path, random_word, basic_playbook_path_w_bv
-
+from ..spellbook import basic_playbook_path, base_image, bad_playbook_path, random_word, basic_playbook_path_w_bv
+from ..conftest import ab
 
 logger = logging.getLogger("ansible_bender")
-
-
-def ab(args, tmpdir_path, return_output=False, ignore_result=False):
-    """
-    python3 -m ab.cli -v build ./playbook.yaml registry.fedoraproject.org/fedora:28 asdqwe-image
-
-    :return:
-    """
-    # put --debug in there for debugging
-    cmd = ["python3", "-m", "ansible_bender.cli", "--database-dir", tmpdir_path] + args
-    logger.debug("cmd = %s", cmd)
-    if ignore_result:
-        return subprocess.call(cmd, cwd=project_dir)
-    if return_output:
-        return subprocess.check_output(cmd, cwd=project_dir, universal_newlines=True)
-    else:
-        # don't use run_cmd here, it makes things complicated
-        subprocess.check_call(cmd, cwd=project_dir)
 
 
 def test_output(target_image, tmpdir):
